adding user and blog functionality done

+++ controller/blogController.js
@@ -1,8 +1,15 @@
-const Blog = require("../models/blogsModel");
+const Blog = require("../models/blogModel");
+const User = require("../models/userModel");
 
 const addBlog = async (req, res) => {
   const formData = req.body;
   try {
+    let user = await User.findById(req.user);
+    if (user.role !== "admin") {
+      return res
+        .status(400)
+        .json({ msg: "Only Admin has access the add Blog" });
+    }
     const data = { ...formData, userId: req.user };
     const blog = new Blog(data);
     const cretedBlog = await blog.save();
@@ -18,8 +25,8 @@ const getBlog = async (req, res) => {
   try {
     const blog = await Blog.findById(blogId);
 
-    if (!blog) {
-      res.status(404).json({ msg: "Blog not Found" });
+    if (!blog) {  
+      return res.status(404).json({ msg: "Blog not Found" });
     }
 
     res.status(200).json({ msg: "Blogs Fetched Successfully", data: blog });
@@ -35,11 +42,18 @@ const updateBlog = async (req, res) => {
     const blog = await Blog.findById(blogId);
 
     if (!blog) {
-      res.status(404).json({ msg: "Blog not Found" });
+      return res.status(404).json({ msg: "Blog not Found" });
+    }
+
+    let user = await User.findById(req.user);
+    if (user.role !== "admin") {
+      return res
+        .status(400)
+        .json({ msg: "Only Admin has access the update Blog" });
     }
 
     if (blog.userId.toString() !== req.user.toString()) {
-      res.status(401).json({ msg: "Only Blog Owner Update the Blog" });
+      return res.status(401).json({ msg: "Only Blog Owner Update the Blog" });
     }
 
     const updetedBlog = await Blog.findByIdAndUpdate(blogId, formData, {
@@ -60,17 +74,24 @@ const deleteBlog = async (req, res) => {
     const blog = await Blog.findById(blogId);
 
     if (!blog) {
-      res.status(404).json({ msg: "Blog not Found" });
+      return res.status(404).json({ msg: "Blog not Found" });
+    }
+
+    let user = await User.findById(req.user);
+    if (user.role !== "admin") {
+      return res
+        .status(400)
+        .json({ msg: "Only Admin has access the update Blog" });
     }
 
     if (blog.userId.toString() !== req.user.toString()) {
-      res.status(401).json({ msg: "Only Blog Owner Delete the Blog" });
+      return res.status(401).json({ msg: "Only Blog Owner Delete the Blog" });
     }
 
     await Blog.findByIdAndDelete(blogId);
 
     res.status(200).json({ msg: "Blog Deleted Successfully" });
-  } catch (error) {
+  } catch (error) { 
     console.log(error);
     res.status(500).json({ msg: "Internal Server Error" });
   }
@@ -94,7 +115,6 @@ const getMyBlogs = async (req, res) => {
     res.status(500).json({ msg: "Internal Server Error" });
   }
 };
-
 
 module.exports = {
   addBlog,
